==> 01_00_P1_Imputation.Rmd <==
---
title: "R Notebook"
output: html_notebook
---

# Missing data and imputation
  
  
TO DO: bring here to this dataset also MMSE, MEMORY AND EF at baseline (check that they are not that highly collinear)
  
  
What type of missing values do we have?
  - most probably missing at random (missing items depend on other variables, e.g. diagnosis, disease stage, etc. but should not depend on the questionnaire itself)
  

Resources on imputation:
https://bookdown.org/mwheymans/bookmi/missing-data-in-questionnaires.html


Based on the reading and the meeting with the statistician:
- include all items per questionnaire but don-t include the total score
- include the total CAV score when imputing FAV items and viceversa
- use predictive mean matching (pmm) for likert items
- The convergence of the imputation should be checked for each imputed item
- After the imputation, the total scores should be computed in each of the imputed datasets and can then be used in the analyses

```{r}

library(tidyverse)
library(here)

set.seed(2023)

```


```{r}
# Load data
dat <- read.csv(here("data","processed", "data_all_atQbaseline.csv"))

dat_colnames <- data.frame(colnames(dat))
```


How much data is missing? 
  - per item
  - how many individuals have at least a missing item? 
  - how many have how much? 


```{r}
n_missing_CAV = sum(dat$CAV_response_rate < 1)
n_missing_FAV = sum(dat$FAV_response_rate < 1)


message(paste("Percentage of individuals with at least one missing item on CAV: ", round(n_missing_CAV/nrow(dat)*100, 2)))
# message(paste("Percentage of individuals with CAV response rate < 0.15: ", round(sum(dat$CAV_response_rate < 0.15)/n_missing_CAV*100, 2)))
# message(paste("Percentage of individuals with CAV response rate < 0.25: ", round(sum(dat$CAV_response_rate < 0.25)/n_missing_CAV*100, 2)))
# message(paste("Percentage of individuals with CAV response rate < 0.5: ", round(sum(dat$CAV_response_rate < 0.5)/n_missing_CAV*100, 2)))
# message(paste("Percentage of individuals with CAV response rate < 0.8: ", round(sum(dat$CAV_response_rate < 0.8)/n_missing_CAV*100, 2)))
message(paste("Percentage of individuals with CAV response rate < 0.15: ", round(sum(dat$CAV_response_rate < 0.15)/nrow(dat)*100, 2)))
message(paste("Percentage of individuals with CAV response rate < 0.25: ", round(sum(dat$CAV_response_rate < 0.25)/nrow(dat)*100, 2)))
message(paste("Percentage of individuals with CAV response rate < 0.5: ", round(sum(dat$CAV_response_rate < 0.5)/nrow(dat)*100, 2)))
message(paste("Percentage of individuals with CAV response rate < 0.75: ", round(sum(dat$CAV_response_rate < 0.75)/nrow(dat)*100, 2)))
message(paste("Percentage of individuals with CAV response rate < 0.9: ", round(sum(dat$CAV_response_rate < 0.9)/nrow(dat)*100, 2)))
message(paste("Percentage of individuals with CAV response rate < 1: ", round(sum(dat$CAV_response_rate < 1)/nrow(dat)*100, 2)))

message("")


message(paste("Percentage of individuals with at least one missing item on FAV: ", round(n_missing_FAV/nrow(dat)*100, 2)))
# message(paste("Percentage of individuals with FAV response rate < 0.15: ", round(sum(dat$FAV_response_rate < 0.15)/n_missing_FAV*100, 2)))
# message(paste("Percentage of individuals with FAV response rate < 0.25: ", round(sum(dat$FAV_response_rate < 0.25)/n_missing_FAV*100, 2)))
# message(paste("Percentage of individuals with FAV response rate < 0.5: ", round(sum(dat$FAV_response_rate < 0.5)/n_missing_FAV*100, 2)))
# message(paste("Percentage of individuals with FAV response rate < 0.8: ", round(sum(dat$FAV_response_rate < 0.8)/n_missing_FAV*100, 2)))
message(paste("Percentage of individuals with FAV response rate < 0.15: ", round(sum(dat$FAV_response_rate < 0.15)/nrow(dat)*100, 2)))
message(paste("Percentage of individuals with FAV response rate < 0.25: ", round(sum(dat$FAV_response_rate < 0.25)/nrow(dat)*100, 2)))
message(paste("Percentage of individuals with FAV response rate < 0.5: ", round(sum(dat$FAV_response_rate < 0.5)/nrow(dat)*100, 2)))
message(paste("Percentage of individuals with FAV response rate < 0.75: ", round(sum(dat$FAV_response_rate < 0.75)/nrow(dat)*100, 2)))
message(paste("Percentage of individuals with FAV response rate < 0.9: ", round(sum(dat$FAV_response_rate < 0.9)/nrow(dat)*100, 2)))
message(paste("Percentage of individuals with FAV response rate < 1: ", round(sum(dat$FAV_response_rate < 1)/nrow(dat)*100, 2)))


```

```{r}
CAV_items <- paste("CAV_", c(1:25),  sep="")

FAV_F_items <- c("FAV_2", "FAV_2A", "FAV_3", "FAV_3B", "FAV_4", "FAV_4B", "FAV_5", "FAV_5B", "FAV_6", 
         "FAV_6B", "FAV_7", "FAV_8", "FAV_9A", "FAV_9B", "FAV_9C", "FAV_9D", "FAV_10", "FAV_10A", "FAV_10B")

FAV_Q_items <- c("Q2", "Q3", "Q4", "Q5", "Q6", "Q7", "Q8", "Q9A", "Q9B", "Q9C", "Q9D", "Q10")
```




# Missing data - print tables

```{r}
CAV_missing <- dat %>%
  select(all_of(c(CAV_items, "CAV_lifetime", "CAV_current", "CAV_past"))) %>% 
  summarise(across(everything(), ~ sum(is.na(.x)))) %>% 
  pivot_longer(everything(),names_to = "Variable", values_to = "N_missing") %>%
  mutate(Total_sample = nrow(dat),
         P_missing = round((N_missing/nrow(dat))*100, 2))


FAV_missing <- dat %>%
  select(all_of(c(FAV_F_items, FAV_Q_items, "FAV_total"))) %>% 
  summarise(across(everything(), ~ sum(is.na(.x)))) %>% 
  pivot_longer(everything(),names_to = "Variable", values_to = "N_missing") %>%
  mutate(Total_sample = nrow(dat),
         P_missing = round((N_missing/nrow(dat))*100, 2))

```



```{r}
hist(dat$CAV_response_rate)
hist(dat$FAV_response_rate)

data.frame(table(dat$CAV_response_rate))
data.frame(table(dat$FAV_response_rate))

plot(ecdf(dat$CAV_response_rate))
plot(ecdf(dat$FAV_response_rate))

```


```{r}
ordered(unique(dat$Q2))
ordered(unique(dat$Q3))
ordered(unique(dat$Q4))
ordered(unique(dat$Q5))
ordered(unique(dat$Q6))
ordered(unique(dat$Q7))
ordered(unique(dat$Q8))
ordered(unique(dat$Q9A))
ordered(unique(dat$Q9B))
ordered(unique(dat$Q9C))
ordered(unique(dat$Q9D))
ordered(unique(dat$Q10))

```



```{r}
# select dataset with variables that will be needed for analyses
dat <- dat %>% 
  select(I_ID, I_sex, I_age_q, I_edu_VE, A_positive, APOE, D_Q_grouped, MMSE_q, N_15WTui, Z_MEM, N_TMTBt, Z_EXEC,
         ADsign_thick_unweighted, wholebrain_thick_weighted, temporoparietal_thick_weighted,
         CAV_1, CAV_2, CAV_3, CAV_4, CAV_5, CAV_6, CAV_7, CAV_8, CAV_9, CAV_10, CAV_11, CAV_12, CAV_13, CAV_14, CAV_15,
         CAV_16, CAV_17, CAV_18, CAV_19, CAV_20, CAV_21, CAV_22, CAV_23, CAV_24, CAV_25, CAV_lifetime, CAV_current, CAV_past,
         Q2, Q3, Q4, Q5, Q6, Q7, Q8, Q9A, Q9B, Q9C, Q9D, Q10, FAV_total)
         
```


```{r}
# Percentage of missing values across all selected variables
(missing_all <- dat %>%
  select(everything()) %>%
  summarise(across(everything(), ~ sum(is.na(.x)))) %>% 
   pivot_longer(everything(),names_to = "Variable", values_to = "N_missing") %>%
  mutate(P_missing = round((N_missing/nrow(dat))*100, 2))
)


```

```{r}
# check distribution of missing MRI across diagnoses
dat <- dat %>% 
  mutate(MRI_missing = if_else(is.na(temporoparietal_thick_weighted), 1, 0), .after=temporoparietal_thick_weighted) %>% 
  mutate(CAV_lifetime_missing = if_else(is.na(CAV_lifetime), 1, 0), .after=CAV_lifetime) %>% 
  mutate(CAV_past_missing = if_else(is.na(CAV_past), 1, 0), .after=CAV_past) %>% 
  mutate(CAV_current_missing = if_else(is.na(CAV_current), 1, 0), .after=CAV_current) %>% 
  mutate(FAV_total_missing = if_else(is.na(FAV_total), 1, 0), .after=FAV_total) %>% 
  mutate(I_edu_VE_missing = if_else(is.na(I_edu_VE), 1, 0), .after=I_edu_VE) %>% 
  mutate(A_positive_missing = if_else(is.na(A_positive), 1, 0), .after=A_positive) %>% 
  mutate(Z_MEM_missing = if_else(is.na(Z_MEM), 1, 0), .after=Z_MEM) %>% 
  mutate(Z_EXEC_missing = if_else(is.na(Z_EXEC), 1, 0), .after=Z_EXEC)


message("MRI proportion of missing across diagnoses")
round(prop.table(table(dat$MRI_missing, dat$D_Q_grouped))*100, 2)

message("CAV_lifetime proportion of missing across diagnoses")
round(prop.table(table(dat$CAV_lifetime_missing, dat$D_Q_grouped))*100, 2)

message("CAV_past proportion of missing across diagnoses")
round(prop.table(table(dat$CAV_past_missing, dat$D_Q_grouped))*100, 2)

message("CAV_current proportion of missing across diagnoses")
round(prop.table(table(dat$CAV_current_missing, dat$D_Q_grouped))*100, 2)

message("FAV proportion of missing across diagnoses")
round(prop.table(table(dat$FAV_total_missing, dat$D_Q_grouped))*100, 2)

message("Edu proportion of missing across diagnoses")
round(prop.table(table(dat$I_edu_VE_missing, dat$D_Q_grouped))*100, 2)

message("A_positive proportion of missing across diagnoses")
round(prop.table(table(dat$A_positive_missing, dat$D_Q_grouped))*100, 2)

message("Z_MEM proportion of missing across diagnoses")
round(prop.table(table(dat$Z_MEM_missing, dat$D_Q_grouped))*100, 2)

message("Z_EXEC proportion of missing across diagnoses")
round(prop.table(table(dat$Z_EXEC_missing, dat$D_Q_grouped))*100, 2)



message("MRI proportion of missing (rows) across missing CAV (cols)")
round(prop.table(table(dat$MRI_missing, dat$CAV_lifetime_missing))*100, 2)


message("MRI proportion of missing (rows) across missing FAV(cols)")
round(prop.table(table(dat$MRI_missing, dat$FAV_total_missing))*100, 2)
```






# Imputation

## Set-up data

```{r}
# select dataset with variables that will be needed for analyses
dat <- dat %>% 
  select(CAV_1, CAV_2, CAV_3, CAV_4, CAV_5, CAV_6, CAV_7, CAV_8, CAV_9, CAV_10, CAV_11, CAV_12, CAV_13, CAV_14, CAV_15,
         CAV_16, CAV_17, CAV_18, CAV_19, CAV_20, CAV_21, CAV_22, CAV_23, CAV_24, CAV_25,
         Q2, Q3, Q4, Q5, Q6, Q7, Q8, Q9A, Q9B, Q9C, Q9D, Q10,
         I_ID, I_sex, I_age_q, I_edu_VE, APOE, D_Q_grouped, 
         MMSE_q, Z_MEM, Z_EXEC, A_positive,
         CAV_lifetime, CAV_current, CAV_past, FAV_total
         )



# set all character variables to factor
dat <- dat %>% 
  mutate(I_sex       = factor(I_sex),
         I_edu_VE    = factor(I_edu_VE, order=T),
         A_positive  = factor(A_positive),
         APOE        = factor(APOE),
         D_Q_grouped = factor(D_Q_grouped)
         )

  
```




## Check correlations


```{r fig.height=8, fig.width=8}
# Correlation matrix for CAV items

GGally::ggcorr(dat[,c(CAV_items, "CAV_lifetime", "CAV_past", "CAV_current")],
               method = c("pairwise.complete.obs", "spearman"),
               nbreaks = 6,
               label = TRUE,
               label_size = 3,
               hjust = 0.85, size=3,
               color = "grey50")


GGally::ggcorr(dat[,FAV_Q_items],
               method = c("pairwise.complete.obs", "spearman"),
               nbreaks = 6,
               label = TRUE,
               label_size = 3,
               hjust = 0.85, size=3,
               color = "grey50")

```



```{r fig.height=8, fig.width=8}
# Correlation matrix for CAV items

dat2 <- dat %>% 
  mutate(I_sex = as.numeric(I_sex),
         APOE = as.numeric(APOE),
         I_edu_VE = as.numeric(I_edu_VE))

GGally::ggcorr(dat2[,c(CAV_items, "CAV_lifetime", "CAV_past", "CAV_current", "FAV_total", "I_sex", "I_age_q", "I_edu_VE", "APOE", "MMSE_q", "Z_MEM", "Z_EXEC")],
               method = c("pairwise.complete.obs", "spearman"),
               nbreaks = 6,
               label = TRUE,
               label_size = 3,
               hjust = 0.85, size=3,
               color = "grey50")


GGally::ggcorr(dat2[,c(FAV_Q_items, "CAV_lifetime", "CAV_past", "CAV_current", "FAV_total", "I_sex", "I_age_q", "I_edu_VE", "APOE", "MMSE_q", "Z_MEM", "Z_EXEC")],
               method = c("pairwise.complete.obs", "spearman"),
               nbreaks = 6,
               label = TRUE,
               label_size = 3,
               hjust = 0.85, size=3,
               color = "grey50")

```



Percentage of incomplete cases  --> i.e. if we were to make a complete observations dataset across all the variables we will use in the analysis, what percentage of observations will we lose?

If I were to do a complete case on the selected dataset - I remain with 2171/4032 = 53% remaining, 47% missing 

```{r}

# dat_complete <- dat[complete.cases(dat),]


# 
# # Percentage of missing if excluding amylois
# dat2 <- dat %>% 
#   select(CAV_1, CAV_2, CAV_3, CAV_4, CAV_5, CAV_6, CAV_7, CAV_8, CAV_9, CAV_10, CAV_11, CAV_12, CAV_13, CAV_14, CAV_15,
#          CAV_16, CAV_17, CAV_18, CAV_19, CAV_20, CAV_21, CAV_22, CAV_23, CAV_24, CAV_25, 
#          Q2, Q3, Q4, Q5, Q6, Q7, Q8, Q9A, Q9B, Q9C, Q9D, Q10,
#          I_ID, I_sex, I_age_q, I_edu_VE, APOE, D_Q_grouped, 
#          MMSE_q, Z_MEM, Z_EXEC)
# 
# dat2_complete <- dat[complete.cases(dat2),]
# 

```



```{r}
dat <- dat %>% 
  mutate(CAV_lifetime_passive = CAV_lifetime,
         CAV_past_passive     = CAV_past,
         CAV_current_passive  = CAV_current,
         FAV_total_passive    = FAV_total)

```



## Imputation

### Passive imputation 


```{r}

# Initiallize model 
imp0 <- mice::mice(dat, maxit = 0)

predQ <- imp0$predictorMatrix
methodQ <- imp0$method

#predQ
methodQ

```

Exclude I_ID all the way

CAV_items:  sex, age, edu, apoe, a_positive, diagnosis, mmse, mem, exec, cav_items, fav_total_passive
FAV_items:  sex, age, edu, apoe, a_positive, diagnosis, mmse, mem, exec, fav_items, cav_past_passive, cav_current_passive

I_Edu_VE :  sex, age,      apoe, a_positive, diagnosis, mmse, mem, exec, fav_total_passive, cav_past_passive, cav_current_passive
APOE :      sex, age, edu,       a_positive, diagnosis, mmse, mem, exec, fav_total_passive, cav_past_passive, cav_current_passive
a_positive: sex, age, edu, apoe,             diagnosis, mmse, mem, exec, fav_total_passive, cav_past_passive, cav_current_passive
MMSE:       sex, age, edu, apoe, a_positive, diagnosis,       mem, exec, fav_total_passive, cav_past_passive, cav_current_passive
MEM:        sex, age, edu, apoe, a_positive, diagnosis, mmse,      exec, fav_total_passive, cav_past_passive, cav_current_passive
EXEC:       sex, age, edu, apoe, a_positive, diagnosis, mmse, mem,       fav_total_passive, cav_past_passive, cav_current_passive


Impute also total scores directly for those that are fully missing them: (I include the total score of the other questionnaire because it's very likely it has it)

FAV_total:    sex, age, edu, apoe, a_positive, diagnosis, mmse, mem, exec, cav_current_passive, cav_current_past
CAV_lifetime: sex, age, edu, apoe, a_positive, diagnosis, mmse, mem, exec, fav_total_passive
CAV_current:  sex, age, edu, apoe, a_positive, diagnosis, mmse, mem, exec, fav_total_passive
CAV_past:     sex, age, edu, apoe, a_positive, diagnosis, mmse, mem, exec, fav_total_passive



```{r}

# define passive imputation of total scores
methodQ["FAV_total_passive"] <- "~I(20*Q2 + 21 *Q3 + 23*(Q4 + Q5) + 30*Q6 + 25*(Q7 + Q8) + 30*Q9A + 36*Q9B + 20*Q9C + 35*Q9D + 21*Q10)"

methodQ["CAV_lifetime_passive"] <- "~I( (CAV_1 + CAV_2 + CAV_3 + CAV_4 + CAV_5 + CAV_6 + CAV_7 + CAV_8 + CAV_9 + CAV_10 + CAV_11 + CAV_12 + CAV_13 + CAV_14 + CAV_15 + CAV_16 + CAV_17 + CAV_18 + CAV_19 + CAV_20 + CAV_21 + CAV_22 + CAV_23 + CAV_24 + CAV_25) /25)"

methodQ["CAV_past_passive"] <- "~I( (CAV_1 + CAV_2 + CAV_3 + CAV_4 + CAV_5 + CAV_6 + CAV_7 + CAV_8 + CAV_9 + CAV_10 + CAV_11 + CAV_12 + CAV_13 + CAV_14 + CAV_15 + CAV_16 + CAV_17 + CAV_18 + CAV_19 + CAV_20) / 20 )"

methodQ["CAV_current_passive"] <- "~I( (CAV_21 + CAV_22 + CAV_23 + CAV_24 + CAV_25) / 5 ) "


# also set CAV_lifetime that will be imputed as total score to be the passive calculation of CAV_past and CAV_current
methodQ["CAV_lifetime"] <- "~I( (20*CAV_past + 5*CAV_current) / 25)"

predQ <- as.matrix(xlsx::read.xlsx(here("P4_imputation_models.xlsx"), sheetIndex = 1, row.names=1))

```


Set: 
m = 40
maxit  = 50 

```{r}
# set visit order
vis <- c("CAV_1", "CAV_2", "CAV_3", "CAV_4", "CAV_5", "CAV_6", "CAV_7", "CAV_8", "CAV_9", "CAV_10", "CAV_11", "CAV_12", "CAV_13", 
         "CAV_14", "CAV_15", "CAV_16", "CAV_17", "CAV_18", "CAV_19", "CAV_20", "CAV_21", "CAV_22", "CAV_23", "CAV_24", "CAV_25", 
         "CAV_lifetime_passive", "CAV_past_passive", "CAV_current_passive",
         "Q2", "Q3", "Q4", "Q5", "Q6", "Q7", "Q8", "Q9A", "Q9B", "Q9C", "Q9D", "Q10", 
         "FAV_total_passive",
         "I_ID", "I_sex", "I_age_q", "I_edu_VE", "APOE", "D_Q_grouped", "MMSE_q", "Z_MEM", "Z_EXEC", "A_positive", 
         "CAV_lifetime", "CAV_current", "CAV_past", "FAV_total")



# imp <- mice::mice(dat,
#                   m = 5, maxit = 50, pred = predQ, method = methodQ, visitSequence = vis,
#                   seed = 2023, printFlag = FALSE)
# 
# #saveRDS(imp, here("data", "processed", "Part1_mids_m5i30.rds"))
# 

 
imp <- mice::mice(dat,
                  m = 40, maxit = 100, pred = predQ, method = methodQ, visitSequence = vis,
                  seed = 2023, printFlag = FALSE)

saveRDS(imp, here("data", "processed", "Part1_mids_m40i100_allin.rds"))


#imp <- readRDS(here("data", "processed", "Part1_mids_m40i50.rds"))

```


#### Checks


? are all vars imputed ok?
? how do the categorical vars look like?

```{r}
# Logged events??
x <- imp$loggedEvents

# Sanity check whether there are still missing values in the imputed datasets
imp_dat <- data.frame(complete(imp, include = TRUE, action = "long"))

imp_dat %>%
  filter(.imp > 0) %>% 
  select(everything()) %>%
  summarise_all(~ sum(is.na(.x))) %>%
  pivot_longer(everything(),names_to = "Variable", values_to = "N_missing")


# If I compute total scores from items, is this the same as the passively imputed variable?  --> yes
# imp_dat$CAV_lifetime_comp <- rowSums(imp_dat[c(paste0("CAV_", 1:25))])
# imp_dat$CAV_past_comp <- rowSums(imp_dat[c(paste0("CAV_", 1:20))])
# imp_dat$CAV_current_comp <- rowSums(imp_dat[c(paste0("CAV_", 21:25))])

```



#### Diagnostics

```{r}
# Check convergence of total scores

plot(imp, c("CAV_lifetime_passive", "CAV_past_passive", "CAV_current_passive", "FAV_total_passive"))

```


```{r}
# Check convergence per items
plot(imp, CAV_items)

```


```{r}
plot(imp, FAV_Q_items)

```

```{r}
# convergence of imputed total scores

plot(imp, c("CAV_lifetime", "CAV_past", "CAV_current", "FAV_total"))

```


```{r}
# convergence of other variables
 plot(imp, c("I_edu_VE", "APOE", "MMSE_q", "Z_MEM", "Z_EXEC", "A_positive"))

```




```{r}
# Plot passively imputed total scores vs. fully imputed total scores
imp_dat %>%
  filter(.imp == 1) %>% 
  ggplot(., aes(x=CAV_lifetime, y=CAV_lifetime_passive)) +
  geom_point()

imp_dat %>%
  filter(.imp == 1) %>% 
  ggplot(., aes(x=CAV_current, y=CAV_current_passive)) +
  geom_point()

imp_dat %>%
  filter(.imp == 1) %>% 
  ggplot(., aes(x=FAV_total, y=FAV_total_passive)) +
  geom_point()
```


```{r}
library(ggridges)

# Plot distributions across imputed datasets (imp=0 should be the complete case distribution)
imp_dat %>% 
  rename(IMP = ".imp") %>% 
  ggplot(., aes(x = CAV_lifetime_passive, y = as.factor(IMP))) +
  geom_density_ridges()

imp_dat %>% 
  rename(IMP = ".imp") %>% 
  ggplot(., aes(x = CAV_past_passive, y = as.factor(IMP))) +
  geom_density_ridges()

imp_dat %>% 
  rename(IMP = ".imp") %>% 
  ggplot(., aes(x = CAV_current_passive, y = as.factor(IMP))) +
  geom_density_ridges()

imp_dat %>% 
  rename(IMP = ".imp") %>% 
  ggplot(., aes(x = FAV_total_passive, y = as.factor(IMP))) +
  geom_density_ridges()

```


```{r}
# Same for the directly imputed total scores

imp_dat %>% 
  rename(IMP = ".imp") %>% 
  ggplot(., aes(x = CAV_lifetime, y = as.factor(IMP))) +
  geom_density_ridges()

imp_dat %>% 
  rename(IMP = ".imp") %>% 
  ggplot(., aes(x = CAV_past, y = as.factor(IMP))) +
  geom_density_ridges()

imp_dat %>% 
  rename(IMP = ".imp") %>% 
  ggplot(., aes(x = CAV_current, y = as.factor(IMP))) +
  geom_density_ridges()

imp_dat %>% 
  rename(IMP = ".imp") %>% 
  ggplot(., aes(x = FAV_total, y = as.factor(IMP))) +
  geom_density_ridges()
```



```{r}
#::stripplot(imp)
# 
# mice::bwplot(imp)
# 
# mice::bwplot(imp, CAV_lifetime_passive ~ .imp)

mice::stripplot(imp, CAV_lifetime_passive ~ .imp)
mice::stripplot(imp, CAV_past_passive ~ .imp)
mice::stripplot(imp, CAV_current_passive ~ .imp)

mice::stripplot(imp, FAV_total_passive ~ .imp)

```
