==> 02_02_P2_Prep_data.Rmd <==
---
title: "R Notebook"
output: html_notebook
---



```{r}

library(tidyverse)
library(here)

```


## Prep original unimputed data

```{r}
# load original unimputed data
dat <- read.csv(here("data","processed", "data_ADcont_atQbaseline.csv"))

# Create edu tertiles and grouped APOE
dat <- dat %>% 
  mutate(I_edu_tert = case_when(I_edu_VE %in% c(1,2,3,4) ~ "Low",
                                I_edu_VE %in% c(5) ~ "Medium",
                                I_edu_VE %in% c(6,7) ~ "High"), .after=I_edu_VE) %>%
  mutate(Edu_years = case_when(I_edu_VE == 1 ~ 5,
                               I_edu_VE == 2 ~ 6,
                               I_edu_VE == 3 ~ 8,
                               I_edu_VE == 4 ~ 9,
                               I_edu_VE == 5 ~ 10,
                               I_edu_VE == 6 ~ 13, 
                               I_edu_VE == 7 ~ 17  ), .after=I_edu_VE) %>% 
  mutate(APOE_grouped = case_when(APOE %in% c("E2E2", "E2E3") ~  "E2 positive",
                                  APOE %in% c("E4E4")         ~  "E4 homozygous",
                                  APOE %in% c("E3E4", "E2E4") ~  "E4 heterozygous",
                                  APOE %in% c("E3E3")         ~  "E4 negative" ), .after = APOE) %>% 
  mutate(APOE_e4 = case_when(APOE %in% c("E2E4","E3E4", "E4E4") ~ "E4 positive",
                             APOE %in% c("E2E2","E3E3", "E2E3") ~ "E4 negative"), .after = APOE)




# Make factors
dat <- dat %>%
  select(I_ID, CAV_date, I_age_q, I_sex, I_edu_VE, Edu_years, I_edu_tert, D_Q_grouped, APOE, APOE_e4, APOE_grouped, A_positive, MMSE_q, Z_MEM, Z_EXEC,
         CAV_lifetime, CAV_past, CAV_current,
         FAV_total, CAV_response_rate, FAV_response_rate, 
         ADsign_thick_unweighted, wholebrain_thick_weighted, temporoparietal_thick_weighted) %>% 
  mutate(I_sex        = factor(I_sex, 
                               levels=c("m","f"), labels=c("Male", "Female")),
         I_edu_tert   = factor(I_edu_tert,
                               levels = c("Low", "Medium", "High")),
         I_edu_VE     = as.numeric(I_edu_VE),
         A_positive   = factor(A_positive),
         APOE         = factor(APOE, levels = c("E2E2", "E2E3", "E2E4", "E3E3", "E3E4", "E4E4")),
         APOE_grouped = factor(APOE_grouped,
                               levels = c("E2 positive", "E4 negative", "E4 heterozygous","E4 homozygous")),
         APOE_e4      = factor(APOE_e4, levels = c("E4 negative", "E4 positive")),
         D_Q_grouped  = factor(D_Q_grouped, 
                               levels=c("SCD", "MCI", "AD"))) %>% 
  rename(Age = I_age_q,
         Sex = I_sex,
         Edu_tert = I_edu_tert,
         Edu_VE = I_edu_VE,
         Diagnosis = D_Q_grouped,
         MMSEq = MMSE_q,
         MEMq = Z_MEM,
         EXECq = Z_EXEC,
         CAV_lifetime = CAV_lifetime,
         CAV_past = CAV_past,
         CAV_current = CAV_current,
         FAV = FAV_total,
         ADsign_thick = ADsign_thick_unweighted,
         WB_thick = wholebrain_thick_weighted,
         TP_thick = temporoparietal_thick_weighted)



# Labelled columns give problems later when plotting with ggeffects, so skip this
# Label columns
dat <- sjlabelled::var_labels(dat, 
                              Sex = "Sex",
                              Age = "Age",
                              Edu_VE	 = "Education (Verhage)",
                              Edu_tert = "Education - tert",
                              Edu_years = "Education - years",
                              CAV_lifetime = "Lifetime CAQ",
                              CAV_past = "Past CAQ",
                              CAV_current = "Current CAQ",
                              FAV = "PASE score",
                              MMSEq = "MMSE bl",
                              MEMq = "Memory bl",
                              EXECq = "Executive function bl",
                              APOE = "APOE genotype",
                              APOE_e4 = "APOE e4 status",
                              APOE_grouped = "APOE genotype",
                              A_positive = "Amyloid status",
                              Diagnosis = "Diagnosis",
                              CAV_response_rate = "CAQ response rate",
                              FAV_response_rate = "PASE response rate",
                              ADsign_thick = "AD signature thickness",
                              WB_thick = "Wholebrain thickness",
                              TP_thick = "Temporoparietal thickness")

```


```{r}

# Create a longitudinal dataset for each cognitive outcome with the unimputed data 

dat_cog <- readxl::read_excel(here("data", "interim","551_cog_reserve_activity_DBo_20231124_Long_Domains_ADcont.xlsx"), col_names = T)


dat_mmse <- dat_cog %>% 
  mutate(MMSE_date = as.Date(V_dat)) %>% 
  select(I_ID, MMSE_date, V_MMSE) %>% 
  drop_na(V_MMSE) %>% 
  left_join(., dat[,c("I_ID", "CAV_date", "Diagnosis", "Age", "Sex", "Edu_VE", "Edu_tert", "Edu_years", "CAV_lifetime", "CAV_past", "CAV_current", "FAV", 
                      "MMSEq", "MEMq", "EXECq", "APOE", "APOE_e4", "APOE_grouped", "A_positive", "CAV_response_rate", 
                      "FAV_response_rate", "ADsign_thick", "WB_thick", "TP_thick")], 
            by="I_ID") %>%
  mutate(Time_days = as.numeric(difftime(MMSE_date, CAV_date, unit="days")),
         Time_years = Time_days/365, .after=CAV_date) %>% 
  filter(Time_years > -2) %>% 
  rename(MMSE = V_MMSE) %>% 
  group_by(I_ID) %>% 
  arrange(I_ID, Time_years) %>%
  mutate(Visit = order(Time_years), .before=MMSE_date) %>%
  ungroup()


dat_mem <- dat_cog %>% 
  mutate(COG_date = as.Date(N_date)) %>% 
  select(I_ID, COG_date, Z_MEM) %>% 
  drop_na(Z_MEM) %>% 
  left_join(., dat[,c("I_ID", "CAV_date", "Diagnosis", "Age", "Sex", "Edu_VE", "Edu_tert", "Edu_years", "CAV_lifetime", "CAV_past", "CAV_current", "FAV", 
                      "MMSEq", "MEMq", "EXECq", "APOE", "APOE_e4", "APOE_grouped", "A_positive", "CAV_response_rate", 
                      "FAV_response_rate", "ADsign_thick", "WB_thick", "TP_thick")], 
            by="I_ID") %>%
  mutate(Time_days = as.numeric(difftime(COG_date, CAV_date, unit="days")),
         Time_years = Time_days/365, .after=CAV_date) %>% 
  filter(Time_years > -2) %>% 
  rename(MEM = Z_MEM) %>% 
  group_by(I_ID) %>% 
  arrange(I_ID, Time_years) %>%
  mutate(Visit = order(Time_years), .before=COG_date) %>%
  ungroup()


dat_exec <- dat_cog %>% 
  mutate(COG_date = as.Date(N_date)) %>% 
  select(I_ID, COG_date, Z_EXEC) %>% 
  drop_na(Z_EXEC) %>% 
  left_join(., dat[,c("I_ID", "CAV_date", "Diagnosis", "Age", "Sex", "Edu_VE", "Edu_tert", "Edu_years", "CAV_lifetime", "CAV_past", "CAV_current", "FAV", 
                      "MMSEq", "MEMq", "EXECq", "APOE", "APOE_e4", "APOE_grouped", "A_positive", "CAV_response_rate", 
                      "FAV_response_rate", "ADsign_thick", "WB_thick", "TP_thick")], 
            by="I_ID") %>%
  mutate(Time_days = as.numeric(difftime(COG_date, CAV_date, unit="days")),
         Time_years = Time_days/365, .after=CAV_date) %>% 
  filter(Time_years > -2) %>% 
  rename(EXEC = Z_EXEC) %>% 
  group_by(I_ID) %>% 
  arrange(I_ID, Time_years) %>%
  mutate(Visit = order(Time_years), .before=COG_date) %>%
  ungroup()


# Z-score longitudinal scores with respect to mean/SD of the visit=0 timepoint
dat_mmse <- dat_mmse %>% 
  mutate(MMSE_s = (MMSE - mean(dat_mmse$MMSE[dat_mmse$Visit == 1], na.rm=TRUE) )/sd(dat_mmse$MMSE[dat_mmse$Visit == 1], na.rm=TRUE), .after=MMSE)

dat_mem <- dat_mem %>% 
  mutate(MEM_s = (MEM - mean(dat_mem$MEM[dat_mem$Visit == 1], na.rm=TRUE) )/sd(dat_mem$MEM[dat_mem$Visit == 1], na.rm=TRUE), .after=MEM)

dat_exec <- dat_exec %>% 
  mutate(EXEC_s = (EXEC - mean(dat_exec$EXEC[dat_exec$Visit == 1], na.rm=TRUE) )/sd(dat_exec$EXEC[dat_exec$Visit == 1], na.rm=TRUE), .after=EXEC)

```


```{r}
# Save RDS object for analyses

# original unimputed cross-sectional dataset
saveRDS(dat, file = here("data", "processed", "Part2_ADcont_dat.rds"))

# original unimputed longitudinal datasets for each outcome
saveRDS(dat_mmse, file = here("data", "processed", "Part2_ADcont_dat_long_mmse.rds"))
saveRDS(dat_mem,  file = here("data", "processed", "Part2_ADcont_dat_long_mem.rds"))
saveRDS(dat_exec, file = here("data", "processed", "Part2_ADcont_dat_long_exec.rds"))

```




## Prep imputed data

The imputed mids object contains:
- CAV items
- FAV items
- Total scores passive
- Total scores directly imputes
- I_ID, age, sex
- Diagnosis, MMSE, MEM, EXEC, A_positive

- Edu VE --> need to tertile this
- APOE --> need to group this


https://nerler.github.io/EP16_Multiple_Imputation/practical/07_Imputation_of_Longitudinal_Data.html   super nice article on imputing data and merging with longitudinal data (they use tho the random effects of longi models on the longi variables as input in the imputation models)

```{r}
# load mids object
datmids <- readRDS(here("data", "processed", "Part2_mids_m40i100.rds"))

# Prep imputed datasets
datimp <- complete(datmids, action="long", include = TRUE)


# Bring from the original datasets CAV_response_rate, FAV_response_rate
datimp <- datimp %>% 
  left_join(., dat[,c("I_ID", "CAV_response_rate", "FAV_response_rate", "ADsign_thick", "WB_thick", "TP_thick")], by="I_ID")


# Calculate final total scores
# If response_rate < 25%, take directly imputed total scores, otherwise take the passively imputed score
datimp <- datimp %>% 
  mutate(CAV_lifetime = if_else(CAV_response_rate < 0.25, CAV_lifetime, CAV_lifetime_passive),
         CAV_past     = if_else(CAV_response_rate < 0.25, CAV_past,     CAV_past_passive),
         CAV_current  = if_else(CAV_response_rate < 0.25, CAV_current,  CAV_current_passive),
         FAV_total    = if_else(FAV_response_rate < 0.25, FAV_total,    FAV_total_passive))


# Calculate edu tertiles and group APOE
datimp <- datimp %>% 
  mutate(I_edu_tert = case_when(I_edu_VE %in% c(1,2,3,4) ~ "Low",
                                 I_edu_VE %in% c(5) ~ "Medium",
                                 I_edu_VE %in% c(6,7) ~ "High"), .after=I_edu_VE) %>% 
   mutate(Edu_years = as.numeric(case_when(I_edu_VE == 1 ~ 5,
                               I_edu_VE == 2 ~ 6,
                               I_edu_VE == 3 ~ 8,
                               I_edu_VE == 4 ~ 9,
                               I_edu_VE == 5 ~ 10,
                               I_edu_VE == 6 ~ 13, 
                               I_edu_VE == 7 ~ 17  )), .after=I_edu_VE) %>% 
  mutate(APOE_grouped = case_when(APOE %in% c("E2E2", "E2E3") ~  "E2 positive",
                                  APOE %in% c("E4E4")         ~  "E4 homozygous",
                                  APOE %in% c("E3E4", "E2E4") ~  "E4 heterozygous",
                                  APOE %in% c("E3E3")         ~  "E4 negative" ), .after = APOE) %>% 
  mutate(APOE_e4 = case_when(APOE %in% c("E2E4","E3E4", "E4E4") ~ "E4 positive",
                             APOE %in% c("E2E2","E3E3", "E2E3") ~ "E4 negative"), .after = APOE)
  

# Make factors
datimp <- datimp %>%
  select('.imp', '.id', I_ID, I_age_q, I_sex, I_edu_VE, I_edu_tert, Edu_years, D_Q_grouped, APOE, APOE_e4, APOE_grouped, 
         MMSE_q, Z_MEM, Z_EXEC, 
         CAV_lifetime, CAV_past, CAV_current,
         FAV_total, ADsign_thick, WB_thick, TP_thick) %>% 
  mutate(I_sex        = factor(I_sex, 
                               levels=c("m","f"), labels=c("Male", "Female")),
         I_edu_tert   = factor(I_edu_tert,
                               levels = c("Low", "Medium", "High")),
         I_edu_VE     = as.numeric(I_edu_VE),
         APOE         = factor(APOE, levels = c("E2E2", "E2E3", "E2E4", "E3E3", "E3E4", "E4E4")),
         APOE_e4      = factor(APOE_e4, levels = c("E4 negative", "E4 positive")),
         APOE_grouped = factor(APOE_grouped,
                               levels = c("E2 positive", "E4 negative", "E4 heterozygous","E4 homozygous")),
         D_Q_grouped  = factor(D_Q_grouped, 
                               levels=c("SCD", "MCI", "AD"))) %>% 
  rename(Age = I_age_q,
         Sex = I_sex,
         Edu_tert = I_edu_tert,
         Edu_VE = I_edu_VE,
         Diagnosis = D_Q_grouped,
         MMSEq = MMSE_q,
         MEMq = Z_MEM,
         EXECq = Z_EXEC,
         CAV_lifetime = CAV_lifetime,
         CAV_past = CAV_past,
         CAV_current = CAV_current,
         FAV = FAV_total)


# center and zscore variables
datimp <- datimp %>%
  group_by(.imp) %>%
  # center variables
  mutate(Age_c          = Age - mean(Age), .after=Age) %>%
  mutate(Edu_years_c    = Edu_years - mean(Edu_years), .after=Edu_years) %>%
  mutate(Edu_VE_c       = Edu_VE - mean(Edu_VE), .after=Edu_VE) %>%
  mutate(CAV_lifetime_c = CAV_lifetime - mean(CAV_lifetime, na.rm=TRUE), .after=CAV_lifetime) %>%
  mutate(CAV_current_c  = CAV_current - mean(CAV_current, na.rm=TRUE), .after=CAV_current) %>%
  mutate(CAV_past_c     = CAV_past - mean(CAV_past, na.rm=TRUE), .after=CAV_past) %>%
  mutate(FAV_c          = FAV - mean(FAV, na.rm=TRUE), .after=FAV) %>%
  mutate(MMSEq_c        = MMSEq - mean(MMSEq, na.rm=TRUE), .after=MMSEq) %>%
  mutate(MEMq_c         = MEMq - mean(MEMq, na.rm=TRUE), .after=MEMq) %>%
  mutate(EXECq_c        = EXECq - mean(EXECq, na.rm=TRUE), .after=EXECq) %>%
  mutate(ADsign_thick_c = ADsign_thick - mean(ADsign_thick, na.rm=TRUE), .after=ADsign_thick) %>%
  mutate(WB_thick_c     = WB_thick - mean(WB_thick, na.rm=TRUE), .after=WB_thick) %>%
  mutate(TP_thick_c     = TP_thick - mean(TP_thick, na.rm=TRUE), .after=TP_thick) %>%


  # scale variables
  mutate(Age_s          = (Age - mean(Age) )/sd(Age), .after=Age) %>%
  mutate(Edu_years_s    = Edu_years - mean(Edu_years, na.rm=TRUE) /sd(Edu_years, na.rm=TRUE), .after=Edu_years) %>%
  mutate(Edu_VE_s       = Edu_VE - mean(Edu_VE, na.rm=TRUE) /sd(Edu_VE, na.rm=TRUE), .after=Edu_VE) %>%
  mutate(CAV_lifetime_s = (CAV_lifetime - mean(CAV_lifetime, na.rm=TRUE) )/sd(CAV_lifetime, na.rm=TRUE), .after=CAV_lifetime) %>%
  mutate(CAV_current_s  = (CAV_current - mean(CAV_current, na.rm=TRUE) )/sd(CAV_current, na.rm=TRUE), .after=CAV_current) %>%
  mutate(CAV_past_s     = (CAV_past - mean(CAV_past, na.rm=TRUE) )/sd(CAV_past, na.rm=TRUE), .after=CAV_past) %>%
  mutate(FAV_s          = (FAV - mean(FAV, na.rm=TRUE) )/sd(FAV, na.rm=TRUE), .after=FAV) %>%
  mutate(MMSEq_s        = (MMSEq - mean(MMSEq, na.rm=TRUE) )/sd(MMSEq, na.rm=TRUE), .after=MMSEq) %>%
  mutate(MEMq_s         = (MEMq - mean(MEMq, na.rm=TRUE) )/sd(MEMq, na.rm=TRUE), .after=MEMq) %>%
  mutate(EXECq_s        = (EXECq - mean(EXECq, na.rm=TRUE) )/sd(EXECq, na.rm=TRUE), .after=EXECq) %>%
  mutate(ADsign_thick_s = (ADsign_thick - mean(ADsign_thick, na.rm=TRUE) )/sd(ADsign_thick, na.rm=TRUE), .after=ADsign_thick) %>%
  mutate(WB_thick_s     = (WB_thick - mean(WB_thick, na.rm=TRUE) )/sd(WB_thick, na.rm=TRUE), .after=WB_thick) %>%
  mutate(TP_thick_s     = (TP_thick - mean(TP_thick, na.rm=TRUE) )/sd(TP_thick, na.rm=TRUE), .after=TP_thick) %>%
  ungroup()


# back to mids
datmids <- mice::as.mids(datimp)

# subset datasets for SCD, MCI, AD and back t mids
datimp_SCD <- datimp %>% filter(Diagnosis == "SCD")
datimp_MCI <- datimp %>% filter(Diagnosis == "MCI")
datimp_AD  <- datimp %>% filter(Diagnosis == "AD")

datmids_SCD <- mice::as.mids(datimp_SCD)
datmids_MCI <- mice::as.mids(datimp_MCI)
datmids_AD  <- mice::as.mids(datimp_AD)


```


```{r}
# Create datasets as list with longi data added to each imputed dataset

datimp <- complete(datmids, action="long", include = FALSE)

datimp_mmse_list <- lapply(split(datimp, datimp$.imp), function(x) {
  merge(x, subset(dat_mmse, select = c(I_ID, Visit, Time_days, Time_years, MMSE, MMSE_s)), all = TRUE) })

datimp_mem_list <- lapply(split(datimp, datimp$.imp), function(x) {
  merge(x, subset(dat_mem, select = c(I_ID, Visit, Time_days, Time_years, MEM, MEM_s)), all = TRUE) })

datimp_exec_list <- lapply(split(datimp, datimp$.imp), function(x) {
  merge(x, subset(dat_exec, select = c(I_ID, Visit, Time_days, Time_years, EXEC, EXEC_s)), all = TRUE) })


# convert datasets back to mids
datmids_mmse <- miceadds::datlist2mids(datimp_mmse_list)
datmids_mem  <- miceadds::datlist2mids(datimp_mem_list)
datmids_exec <- miceadds::datlist2mids(datimp_exec_list)

```


```{r}
saveRDS(datmids, file = here("data", "processed", "Part2_ADcont_datmids.rds"))

saveRDS(datmids_SCD, file = here("data", "processed", "Part2_ADcont_datmids_SCD.rds"))
saveRDS(datmids_MCI, file = here("data", "processed", "Part2_ADcont_datmids_MCI.rds"))
saveRDS(datmids_AD,  file = here("data", "processed", "Part2_ADcont_datmids_AD.rds"))

saveRDS(datmids_mmse, file = here("data", "processed", "Part2_ADcont_datmids_mmse.rds"))
saveRDS(datmids_mem,  file = here("data", "processed", "Part2_ADcont_datmids_mem.rds"))
saveRDS(datmids_exec, file = here("data", "processed", "Part2_ADcont_datmids_exec.rds"))

```


```{r}
# Make and save strata mids for each diagnosis

# ------ SCD -------
# ------------------
datimp_mmse_SCD_list <- lapply(split(datimp, datimp$.imp), function(x) {
  merge(x, subset(dat_mmse, select = c(I_ID, Visit, Time_days, Time_years, MMSE, MMSE_s)),  all = TRUE) %>% 
  filter(Diagnosis == "SCD") })

datimp_mem_SCD_list <- lapply(split(datimp, datimp$.imp), function(x) {
  merge(x, subset(dat_mem, select = c(I_ID, Visit, Time_days, Time_years, MEM, MEM_s)), all = TRUE) %>% 
  filter(Diagnosis == "SCD") })

datimp_exec_SCD_list <- lapply(split(datimp, datimp$.imp), function(x) {
  merge(x, subset(dat_exec, select = c(I_ID, Visit, Time_days, Time_years, EXEC, EXEC_s)), all = TRUE) %>% 
  filter(Diagnosis == "SCD") })

# convert datasets back to mids
datmids_mmse_SCD <- miceadds::datlist2mids(datimp_mmse_SCD_list)
datmids_mem_SCD  <- miceadds::datlist2mids(datimp_mem_SCD_list)
datmids_exec_SCD <- miceadds::datlist2mids(datimp_exec_SCD_list)


# ------ MCI -------
# ------------------
datimp_mmse_MCI_list <- lapply(split(datimp, datimp$.imp), function(x) {
  merge(x, subset(dat_mmse, select = c(I_ID, Visit, Time_days, Time_years, MMSE, MMSE_s)),  all = TRUE) %>% 
  filter(Diagnosis == "MCI") })

datimp_mem_MCI_list <- lapply(split(datimp, datimp$.imp), function(x) {
  merge(x, subset(dat_mem, select = c(I_ID, Visit, Time_days, Time_years, MEM, MEM_s)), all = TRUE) %>% 
  filter(Diagnosis == "MCI") })

datimp_exec_MCI_list <- lapply(split(datimp, datimp$.imp), function(x) {
  merge(x, subset(dat_exec, select = c(I_ID, Visit, Time_days, Time_years, EXEC, EXEC_s)), all = TRUE) %>% 
  filter(Diagnosis == "MCI") })

# convert datasets back to mids
datmids_mmse_MCI <- miceadds::datlist2mids(datimp_mmse_MCI_list)
datmids_mem_MCI  <- miceadds::datlist2mids(datimp_mem_MCI_list)
datmids_exec_MCI <- miceadds::datlist2mids(datimp_exec_MCI_list)



# ------ AD --------
# ------------------
datimp_mmse_AD_list <- lapply(split(datimp, datimp$.imp), function(x) {
  merge(x, subset(dat_mmse, select = c(I_ID, Visit, Time_days, Time_years, MMSE, MMSE_s)),  all = TRUE) %>% 
  filter(Diagnosis == "AD") })

datimp_mem_AD_list <- lapply(split(datimp, datimp$.imp), function(x) {
  merge(x, subset(dat_mem, select = c(I_ID, Visit, Time_days, Time_years, MEM, MEM_s)), all = TRUE) %>% 
  filter(Diagnosis == "AD") })

datimp_exec_AD_list <- lapply(split(datimp, datimp$.imp), function(x) {
  merge(x, subset(dat_exec, select = c(I_ID, Visit, Time_days, Time_years, EXEC, EXEC_s)), all = TRUE) %>% 
  filter(Diagnosis == "AD") })

# convert datasets back to mids
datmids_mmse_AD <- miceadds::datlist2mids(datimp_mmse_AD_list)
datmids_mem_AD  <- miceadds::datlist2mids(datimp_mem_AD_list)
datmids_exec_AD <- miceadds::datlist2mids(datimp_exec_AD_list)

```



```{r}
saveRDS(datmids_mmse_SCD, file = here("data", "processed", "Part2_ADcont_SCD_datmids_mmse.rds"))
saveRDS(datmids_mem_SCD,  file = here("data", "processed", "Part2_ADcont_SCD_datmids_mem.rds"))
saveRDS(datmids_exec_SCD, file = here("data", "processed", "Part2_ADcont_SCD_datmids_exec.rds"))

saveRDS(datmids_mmse_MCI, file = here("data", "processed", "Part2_ADcont_MCI_datmids_mmse.rds"))
saveRDS(datmids_mem_MCI,  file = here("data", "processed", "Part2_ADcont_MCI_datmids_mem.rds"))
saveRDS(datmids_exec_MCI, file = here("data", "processed", "Part2_ADcont_MCI_datmids_exec.rds"))

saveRDS(datmids_mmse_AD, file = here("data", "processed", "Part2_ADcont_AD_datmids_mmse.rds"))
saveRDS(datmids_mem_AD,  file = here("data", "processed", "Part2_ADcont_AD_datmids_mem.rds"))
saveRDS(datmids_exec_AD, file = here("data", "processed", "Part2_ADcont_AD_datmids_exec.rds"))

```




# Prep data for plotting

Take the median at the individual-level across the multiple imputed datasets for plotting purposed. 

```{r}

# Take the imputed sets only
datimp <- complete(datmids, action="long", include = FALSE)

# make one dataset with the median value across the imputed variables
dat_median <- datimp %>% 
  group_by(.id) %>% 
  select(.id, I_ID, Age, Edu_years, MMSEq, MEMq, EXECq, CAV_lifetime, CAV_past, CAV_current, FAV, ADsign_thick, WB_thick, TP_thick) %>% 
  summarise_all(.funs = median)


# declare function to extract mode
calculate_mode <- function(x) {
  uniqx <- unique(x)
  uniqx[which.max(tabulate(match(x, uniqx)))] }


dat_mode <- datimp %>% 
  group_by(.id) %>% 
  select(.id, Sex, Edu_VE, Edu_tert, Diagnosis, APOE_e4, APOE, APOE_grouped) %>% 
  summarise_all(.funs = calculate_mode)


dat_median <- dat_median %>% 
  left_join(., dat_mode, by=".id")




# Center/scale variables

dat_median <- dat_median %>%
  # scale variables
  mutate(Age_s          = (Age - mean(Age) )/sd(Age), .after=Age) %>%
  mutate(Edu_years_s    = Edu_years - mean(Edu_years, na.rm=TRUE) /sd(Edu_years, na.rm=TRUE), .after=Edu_years) %>%
  mutate(Edu_VE_s       = Edu_VE - mean(Edu_VE, na.rm=TRUE) /sd(Edu_VE, na.rm=TRUE), .after=Edu_VE) %>%
  mutate(CAV_lifetime_s = (CAV_lifetime - mean(CAV_lifetime, na.rm=TRUE) )/sd(CAV_lifetime, na.rm=TRUE), .after=CAV_lifetime) %>%
  mutate(CAV_current_s  = (CAV_current - mean(CAV_current, na.rm=TRUE) )/sd(CAV_current, na.rm=TRUE), .after=CAV_current) %>%
  mutate(CAV_past_s     = (CAV_past - mean(CAV_past, na.rm=TRUE) )/sd(CAV_past, na.rm=TRUE), .after=CAV_past) %>%
  mutate(FAV_s          = (FAV - mean(FAV, na.rm=TRUE) )/sd(FAV, na.rm=TRUE), .after=FAV) %>%
  mutate(MMSEq_s         = (MMSEq - mean(MMSEq, na.rm=TRUE) )/sd(MMSEq, na.rm=TRUE), .after=MMSEq) %>%
  mutate(MEMq_s          = (MEMq - mean(MEMq, na.rm=TRUE) )/sd(MEMq, na.rm=TRUE), .after=MEMq) %>%
  mutate(EXECq_s         = (EXECq - mean(EXECq, na.rm=TRUE) )/sd(EXECq, na.rm=TRUE), .after=EXECq) %>%
  mutate(ADsign_thick_s = (ADsign_thick - mean(ADsign_thick, na.rm=TRUE) )/sd(ADsign_thick, na.rm=TRUE), .after=ADsign_thick) %>%
  mutate(WB_thick_s     = (WB_thick - mean(WB_thick, na.rm=TRUE) )/sd(WB_thick, na.rm=TRUE), .after=WB_thick) %>%
  mutate(TP_thick_s     = (TP_thick - mean(TP_thick, na.rm=TRUE) )/sd(TP_thick, na.rm=TRUE), .after=TP_thick)

# # Label columns
# dat_median <- sjlabelled::var_labels(dat_median,
#                                      Sex = "Sex",
#                                      Age = "Age",
#                                      Age_s = "Age",
#                                      Edu_VE	 = "Education",
#                                      Edu_tert = "Education - tert",
#                                      CAV_lifetime = "Lifetime CAQ",
#                                      CAV_lifetime_s = "Lifetime CAQ",
#                                      CAV_past = "Past CAQ",
#                                      CAV_past_s = "Past CAQ",
#                                      CAV_current = "Current CAQ",
#                                      CAV_current_s = "Current CAQ",
#                                      FAV = "PASE score",
#                                      FAV_s = "PASE score",
#                                      MMSEq = "MMSE",
#                                      MMSEq_s = "MMSE",
#                                      MEMq = "Memory",
#                                      MEMq_s = "Memory",
#                                      EXECq = "Executive function",
#                                      EXECq_s = "Executive function",
#                                      APOE = "APOE genotype",
#                                      APOE_e4 = "APOE e4 status",
#                                      APOE_grouped = "APOE genotype",
#                                      Diagnosis = "Diagnosis",
#                                      ADsign_thick = "AD signature thickness",
#                                      WB_thick = "Wholebrain thickness",
#                                      TP_thick = "Temporoparietal thickness"
# )

```

```{r}
# Construct a longitudinal dataset for each outcome with the median of imputed datasets

dat_median_mmse <- dat_mmse %>% 
  select(I_ID, Visit, Time_days, Time_years, MMSE, MMSE_s) %>% 
  left_join(., dat_median, by="I_ID")

dat_median_mem <- dat_mem %>% 
  select(I_ID, Visit, Time_days, Time_years, MEM, MEM_s) %>% 
  left_join(., dat_median, by="I_ID")

dat_median_exec <- dat_exec %>%
  select(I_ID, Visit, Time_days, Time_years, EXEC, EXEC_s) %>% 
  left_join(., dat_median, by="I_ID")

```


```{r}

saveRDS(dat_median, file = here("data", "processed", "Part2_ADcont_dat_median.rds"))

saveRDS(dat_median_mmse, file = here("data", "processed", "Part2_ADcont_median_mmse.rds"))
saveRDS(dat_median_mem,  file = here("data", "processed", "Part2_ADcont_median_mem.rds"))
saveRDS(dat_median_exec, file = here("data", "processed", "Part2_ADcont_median_exec.rds"))

```
